Validate Response Object
When users are beginning to use the api, or forget the correct options
to pass, there was no clear indication to the user that they were doing
so. This resulted in mockyeah still sending `200`s and just showing a
blank response.

We should be able to inform the user that they did not correctly pass
the right options, and subsequently break execution.